All map layers extracted to separate files
