+++ src/data/CircularEcon.jsx
@@ -1,39 +1,3 @@
-import ReactDOMServer from "react-dom/server";
-import { point, marker, divIcon } from 'leaflet';
-import { geoDateToLocaleString, getStyleMapProperty, mapPopupContent, ogmIconLink, pointToIcon } from './LeafletStyleHelpers';
-// GeoJSON
-import RecyclingDepots from './geojson/RecyclingDepots.geojson';
-import HartlandLandfill from './geojson/HartlandPie.geojson';
-import SingleUsePlastics from './geojson/SingleUsePlastics.geojson';
-// SVG
-import { ReactComponent as RecyclingIcon } from './svg/Recycling.svg';
-import { ReactComponent as HartlandPie } from './svg/HartlandPie.svg';
-
-const styleMap_SUP = new Map([
-  ["None", {fillColor: 'rgb(90,90,90)', color: 'rgb(130,130,130)', legendText: 'No Bylaws'}],
-  ["Bags", {fillColor: 'rgb(0,120,120)', color: 'rgb(130,130,130)', legendText: 'Checkout Bags'}],
-  ["BSC", {fillColor: 'rgb(120,0,120)', color: 'rgb(130,130,130)', legendText: 'Bags, Straws, Containers'}]
-]);
-
-const RecyclingIconStyle = {
-  fill: 'rgb(15,86,229)',
-  icon: <RecyclingIcon />
-}
-
-const styleMap_CircEconBusinesses = new Map([
-  ["5ca7bfb8ecd8490100cab973",{ src: ogmIconLink("5ca7bfb8ecd8490100cab973"), legendText: 'Reuse Shop or Market' }],
-  ["5ca7bfbcecd8490100cab976",{ src: ogmIconLink("5ca7bfbcecd8490100cab976"), legendText: 'Local Business' }],
-  ["5ca7bfbdecd8490100cab979",{ src: ogmIconLink("5ca7bfbdecd8490100cab979"), legendText: 'Green Enterprise' }],
-  ["5ca7bfbeecd8490100cab97b",{ src: ogmIconLink("5ca7bfbeecd8490100cab97b"), legendText: 'Reuse' }],
-  ["5ca7bfc0ecd8490100cab981",{ src: ogmIconLink("5ca7bfc0ecd8490100cab981"), legendText: 'Recycling' }],
-  ["5ca7bfdfecd8490100cab9d1",{ src: ogmIconLink("5ca7bfdfecd8490100cab9d1"), legendText: 'Environmental Education' }],
-  ["5ca7bfe6ecd8490100cab9e6",{ src: ogmIconLink("5ca7bfe6ecd8490100cab9e6"), legendText: 'Artisan or Art Studio' }],
-  ["6067a8ada4eddf0100334b64",{ src: ogmIconLink("6067a8ada4eddf0100334b64"), legendText: 'Farmers/Local Market' }],
-  ["6067a9efa4eddf0100334b68",{ src: ogmIconLink("6067a9efa4eddf0100334b68"), legendText: 'Artisanal Food /Local Flavor' }],
-  ["6067aafab709040100c13c98",{ src: ogmIconLink("6067aafab709040100c13c98"), legendText: 'Composting Site' }],
-  ["5c312ab437407e1f05ac394f",{ src: ogmIconLink("5c312ab437407e1f05ac394f"), legendText: 'SDG 12 - Responsible Consumption and Production' }],
-]);
-
 const CircularEcon = {
   title: 'Circular Economy?',
   question: 'How can I close the loop on waste?',
@@ -113,157 +77,5 @@ const CircularEcon = {
       },
     ],
   },
-  mapData: [
-    {
-      title: 'Single-use Item Bylaws',
-      description: [
-        {type:'p', content:'Single-use items are products designed to be used once and then thrown away. Examples include checkout bags, straws, stir-sticks, and take-away food containers.'},
-        {type: 'p', content: 'Every day, Victoria residents throw away over 75,000 single-use items.¹'},
-        {type:'p', content:'Click the map within a given municipality to learn about their bylaws, or the B.C. Government keeps a comprehensive list of municipal bylaws at the link below:'},
-        {
-          type:'link',
-          content:'Ministry of Environment and Climate Change. Municipal Bylaws for Single-Use Plastics - Province of British Columbia.',
-          url:'https://www2.gov.bc.ca/gov/content/environment/waste-management/zero-waste/municipal-plastics-bylaws#participating-municipalities'
-        },
-        {
-          type:'link',
-          content:'1. City of Victoria. "Single-use Items." Accessed May 26, 2023.',
-          url:'https://www.victoria.ca/EN/main/residents/waste-reduction/single-use-items.html'
-        }
-      ],
-      data: SingleUsePlastics,
-      format: 'polygon',
-      symbology: 'classified',
-      styleMap: styleMap_SUP,
-      options: {
-        style: function (feature) {
-          const baseStyle = {
-            stroke: true,
-            opacity: 0.5,
-            color: 'rgb(130,130,130)',
-            dashArray: '',
-            lineCap: 'butt',
-            lineJoin: 'miter',
-            weight: 3,
-            fill: true,
-            fillOpacity: 0.2,
-            interactive: true
-          }
-          return {
-            ...baseStyle,
-            fillColor: getStyleMapProperty(
-                'fillColor',
-                feature.properties.ImplementationDate === null
-                  ? "None"
-                  : feature.properties.StrawBan === false
-                    ? "Bags"
-                    : "BSC",
-                styleMap_SUP
-              ),
-          }
-        },
-        onEachFeature: (f,l) => {
-          l.bindPopup(mapPopupContent(
-              f.properties.LocalGovShort,
-              f.properties.ImplementationDate === null
-                ? `There are no single-use items bylaws currently in effect in the ${f.properties.LocalGov}.`
-                : `The ${
-                    f.properties.LocalGov
-                  } bylaw for single-use items took effect on ${
-                    geoDateToLocaleString(f.properties.ImplementationDate)
-                  }. Paper checkout bags must cost $${
-                    f.properties.PaperBagFee.toFixed(2)
-                  } and reusable bags must cost $${
-                    f.properties.ReusableBagFee.toFixed(2)
-                  }. Single-use straws and food containers are ${
-                    f.properties.StrawBan ? 'also' : 'not'
-                  } banned. Read the bylaw here:`,
-                  f.properties.URL
-            ), {offset: point(0,8)});
-          l.on({
-            mouseover: (e) => e.target.setStyle({ fillOpacity: 0.7 }),
-            mouseout: (e) => e.target.setStyle({ fillOpacity: 0.2 })
-          });
-        }
-      }
-    },
-    {
-      title: 'Circular Economy Buisinesses',
-      description: [
-        {type: 'p', content: "This is a collection of reuse stores and other businesses engaged in the circular economy through reuse, sustainable practices, local food resources, and green education."},
-        {type: 'link', content: 'Visit and contribute to this project at OpenGreenMap.', url: 'https://new.opengreenmap.org/browse/sites?map=6474bbf4261bc10100e28d90'},
-      ],
-      data: require('./geojson/CircularEconBusinesses.geojson'),
-      shape: 'point',
-      symbology: 'classified',
-      styleMap: styleMap_CircEconBusinesses,
-      options: {
-        pointToLayer: (f,l) => pointToIcon(l, {icon: <img
-            width="26px"
-            height="26px"
-            src={`https://new.opengreenmap.org/api-v1/icons/${f.properties.icons[0]}/image/value`}
-            alt={styleMap_CircEconBusinesses.get(f.properties.icons[0]).legendText}
-          />}),
-        onEachFeature: (f,l) => {
-          const imageURL = f.properties.pictures[0]
-            ? `https://new.opengreenmap.org/api-v1/pictures/${f.properties.pictures[0]}/picture/sm`
-            : null;
-          
-          l.bindPopup(mapPopupContent(
-              f.properties.name,
-              f.properties.description[1].content,
-              `https://new.opengreenmap.org/browse/sites/${f.properties.id}`,
-              'Show More at OpenGreenMap.org',
-              imageURL
-            ), {offset: point(0,-6)});
-        }
-      },
-    },
-    {
-      title: 'Recycling Depots',
-      description: 'Find where you can recycle a wide range of household materials in your neighbourhood. Visit the website of the specific depot to confirm what is accepted there by clicking on it.',
-      data: RecyclingDepots,
-      shape: 'point',
-      symbology: 'single',
-      icon: RecyclingIconStyle,
-      options: {
-        pointToLayer: (f,l) => pointToIcon(l, RecyclingIconStyle),
-        onEachFeature: (f,l) => {
-          l.bindPopup(mapPopupContent(
-            f.properties['Name'],
-            f.properties['Address'],
-            f.properties['URL'],
-            f.properties['URL']
-            ), {offset: point(0,-6)});
-        }
-      }
-    },
-    {
-      title: 'Hartland Landfill',
-      description: 'Overall Waste Composition 2022',
-      data: HartlandLandfill,
-      shape: 'point',
-      symbology: 'single',
-      icon: { icon: <HartlandPie /> },
-      options: {
-        pointToLayer: (f,l) => {
-          return marker(l, {
-            icon: divIcon({
-              className: "",
-              iconSize: [256, 256],
-              iconAnchor: [128, 128],
-              html: ReactDOMServer.renderToString(<HartlandPie />),
-            })
-          })
-        },
-        onEachFeature: (f,l) => {
-          l.bindPopup(mapPopupContent(
-            f.properties['Name'],
-            <HartlandPie />,
-            ), {offset: point(0,-72)});
-        }
-      }
-    },
-  ],
 };
 export default CircularEcon;
