+++ src/data/Flooding.jsx
@@ -1,68 +1,3 @@
-import { icon, marker, point } from 'leaflet';
-import { getStyleMapProperty, mapPopupContent, pointToIcon } from './LeafletStyleHelpers';
-// SVG
-
-const styleMap_RainGardens = new Map([
-  ['641b6ff089e06d0100e0cf1d', {
-    src: `https://new.opengreenmap.org/api-v1/icons/641b6ff089e06d0100e0cf1d/image/value`,
-    legendText: 'Rain Garden'
-  }],
-  ['641b716b89e06d0100e0cf2c', {
-    src: `https://new.opengreenmap.org/api-v1/icons/641b716b89e06d0100e0cf2c/image/value`,
-    legendText: 'Bioswale'
-  }],
-  ['641b714889e06d0100e0cf24', {
-    src: `https://new.opengreenmap.org/api-v1/icons/641b714889e06d0100e0cf24/image/value`,
-    legendText: 'Curb Cut'
-  }],
-]);
-
-const styleMap_RestorationSites = new Map([
-  ['5ebee86cd060db010045ee3c', {
-    src: `https://new.opengreenmap.org/api-v1/icons/5ebee86cd060db010045ee3c/image/value`,
-    legendText: 'Protected / Restored Habitat'
-  }],
-  ['5ca7bff1ecd8490100caba0b', {
-    src: `https://new.opengreenmap.org/api-v1/icons/5ca7bff1ecd8490100caba0b/image/value`,
-    legendText: 'Aquatic Habitat'
-  }],
-  ['5ca7bfeaecd8490100cab9f3', {
-    src: `https://new.opengreenmap.org/api-v1/icons/5ca7bfeaecd8490100cab9f3/image/value`,
-    legendText: 'Waterfront or Riverside Park'
-  }],
-  ['5ca7bfe8ecd8490100cab9ee', {
-    src: `https://new.opengreenmap.org/api-v1/icons/5ca7bfe8ecd8490100cab9ee/image/value`,
-    legendText: 'Water Feature'
-  }],
-  ['5ca7bff4ecd8490100caba0f', {
-    src: `https://new.opengreenmap.org/api-v1/icons/5ca7bff4ecd8490100caba0f/image/value`,
-    legendText: 'Significant Habitat'
-  }],
-  ['5ca7bfe7ecd8490100cab9ec', {
-    src: `https://new.opengreenmap.org/api-v1/icons/5ca7bfe7ecd8490100cab9ec/image/value`,
-    legendText: 'Wetlands'
-  }],
-  ['5ca7bfc1ecd8490100cab983', {
-    src: `https://new.opengreenmap.org/api-v1/icons/5ca7bfc1ecd8490100cab983/image/value`,
-    legendText: 'Greenhouse Gas Reduction'
-  }],
-  ['5f036d59352b920100f4b7f1', {
-    src: `https://new.opengreenmap.org/api-v1/icons/5f036d59352b920100f4b7f1/image/value`,
-    legendText: 'People of Place - Indigenous Culture'
-  }],
-  ['5ca7bfececd8490100cab9fc', {
-    src: `https://new.opengreenmap.org/api-v1/icons/5ca7bfececd8490100cab9fc/image/value`,
-    legendText: 'Canoe or Kayak Site'
-  }],
-]);
-
-const styleMap_CRDInundation = new Map([
-  ["0",   {legendText: '0.0 m', color: '#1E44D3'}],
-  ["0.5", {legendText: '0.5 m', color: '#5177E1'}],
-  ["1",   {legendText: '1.0 m', color: '#81A4EE'}],
-  ["2",   {legendText: '2.0 m', color: '#C6DBFA'}]
-]);
-
 const Flooding = {
   title: 'Protect from Flooding?',
   question: 'How can communities manage water for maximum benefit?',
@@ -119,150 +54,6 @@ const Flooding = {
         description: 'The 1000 Rain Gardens Project: "Making Bowker Creek healthy again requires that people send much more rainwater into the earth and much less into the storm drains. Wherever we live, work or go to school in Bowker Creek’s valley, the 1000 Rain Gardens Project aims to help us manage rainwater better. The project partners Friends of Bowker Creek Society with Peninsula Streams Society (PSS). The PSS initiative, the Rain Gardens in Headwaters Program, will multiply rain gardens throughout Greater Victoria, starting here in Bowker Creek’s valley. The 1000 Rain Gardens project provides a vital first step and proving ground for the wider program."'
       }
     ],
-  },
-  mapData: [
-    {
-      title: 'Watersheds',
-      description: [
-        {type: 'p', content: 'A watershed (also known as a hydrologic unit, catchment, or drainage basin) is defined as that area of land draining into a particular stream or other surface water body. For any location in a stream, there is an associated area that contributes water to its flow. The watershed divide is that line which divides the area contributing water to the stream and that which contributes water to neighboring streams or water bodies. Therefore, each watershed is defined by its outlet or pour point (the point in the stream which receives all water in the watershed) and the associated watershed divide derived from that point and the local topography.¹'},
-        {type: 'p', content: 'If a drop of rain or snow falls on any spot within the boundary of a watershed, if it is not absorbed into the ground or somehow otherwise collected, it can be expected to eventually make its way to the central stream and flow along it to the sea. Even water that soaks into the ground can flow towards the stream as it reaches denser soils or bedrock that it cannot infiltrate. Some water travels much deeper into underground resevoirs, called aquifers, that provide drinking water when accessed by a well.²'},
-        {type: 'link', content: '1. GeoBC, 2010. "Freshwater Atlas User Guide." Integrated Land Management Bureau, Victoria, BC. ', url: 'https://www2.gov.bc.ca/assets/gov/data/geographic/topography/fwa/fwa_user_guide.pdf'},
-        {type: 'link', content: '2. NOAA. "What is a watershed?" Accessed: May 15, 2023.', url: 'https://oceanservice.noaa.gov/facts/watershed.html'},
-      ],
-      data: require('./geojson/CRDWatersheds.geojson'),
-      shape: 'polygon',
-      symbology: 'single',
-      options: {
-        style: {
-          stroke: true,
-          opacity: 0.7,
-          color: 'rgb(28,163,236)',
-          dashArray: '',
-          lineCap: 'butt',
-          lineJoin: 'miter',
-          weight: 3,
-          fill: true,
-          fillOpacity: 0,
-          fillColor: 'rgb(132,182,185)',
-          interactive: true,
-        },
-        onEachFeature: (f,l) => {
-          l.bindPopup(
-            mapPopupContent(
-              f.properties.Name.charAt(0) === '9' ? 'Watershed' : f.properties.Name + ' Watershed',
-              f.properties.Name.charAt(0) === '9'
-                ? `This watershed is not associated with a named stream by GeoBC. Its Freshwater Atlas Code is ${f.properties.Name}. Its area is ${f.properties.Hectares} hectares.`
-                : f.properties.Name + ' Watershed encloses an area of ' + f.properties.Hectares + ' hectares.',
-              ),
-            {offset: point(0,8)});
-          l.on({
-            mouseover: (e) => e.target.setStyle({ fillOpacity: 0.7 }),
-            mouseout: (e) => e.target.setStyle({ fillOpacity: 0 })
-          });
-        }
-      }
-    },
-    {
-      title: 'Coastal Flood Inundation Scenarios',
-      description: [
-        {type: 'p', content: 'The Capital Regional District (CRD) retained Associated Engineering, DHI and Westmar Advisors to undertake the Capital Region Coastal Flood Inundation Mapping Project (the project) in 2019 -2021. On behalf of local government and other regional stakeholders, staff worked closely with an inter-municipal and multi-disciplinary project team to complete the Regional Coastal Flood Inundation Project in order to better understand regional impacts from coastal storm flooding due to sea level rise and tsunamis. Modeled scenarios show the higher high water large tide (HHWLT) levels for a 0.0m, 0.5m, 1.0m, and 2.0m relative sea level rise (RSLR), respectively.'},
-        {type: 'link', content: 'Capital Regional District, 2021. Capital Region Coastal Flood Inundation Mapping Project Summary, Version 2.0. Victoria.', url: 'https://www.crd.bc.ca/docs/default-source/climate-action-pdf/reports/2020-sea-level-mapping-project/coastal-flood-inundation-mapping-project-summary.pdf'},
-      ],
-      data: require('./geojson/CRDInundation.geojson'),
-      shape: 'line',
-      symbology: 'classified',
-      styleMap: styleMap_CRDInundation,
-      options: {
-        style: function (feature) {
-          const baseStyle = {
-            stroke: true,
-            fill: false,
-            interactive: true,
-            weight: 4,
-          }
-          return {
-            ...baseStyle,
-            color: getStyleMapProperty(
-              'color',
-              feature.properties.RSLR,
-              styleMap_CRDInundation
-            ),
-          }
-        },
-        onEachFeature: (f,l) => {
-          l.bindPopup(
-            mapPopupContent(
-              'Capital Region Coastal Flood Inundation',
-              'Scenario of ' + f.properties.RSLR + 'm Relative Sea Level Rise (RSLR), this line represents the Higher High Water Large Tide (HHWLT) level.',
-              ),
-            {offset: point(0,8)});
-        }
-      }
-    },
-    {
-      title: 'Restoration Sites',
-      description: [
-        {type: 'p', content: "Mapping restoration sites as described by Peninsula Streams Society for the RUSH Initiative."},
-        {type: 'link', content: 'Visit and contribute to this project at OpenGreenMap.', url: 'https://new.opengreenmap.org/browse/sites?map=644acbf92af3a50100a0f0bc'},
-      ],
-      data: require('./geojson/RestorationSites.geojson'),
-      shape: 'point',
-      symbology: 'classified',
-      styleMap: styleMap_RestorationSites,
-      options: {
-        pointToLayer: (f,l) => pointToIcon(l, {icon: <img 
-            src={`https://new.opengreenmap.org/api-v1/icons/${f.properties.icons[0]}/image/value`}
-            alt={styleMap_RestorationSites.get(f.properties.icons[0]).legendText}
-          />}),
-        onEachFeature: (f,l) => {
-          const imageURL = f.properties.pictures[0]
-            ? `https://new.opengreenmap.org/api-v1/pictures/${f.properties.pictures[0]}/picture/sm`
-            : null;
-          
-          l.bindPopup(mapPopupContent(
-              f.properties.name,
-              '',
-              `https://new.opengreenmap.org/browse/sites/${f.properties._id}`,
-              'Show More at OpenGreenMap.org',
-              imageURL
-            ), {offset: point(0,-6)});
-        }
-      },
-    },
-    {
-      title: '1000 Rain Gardens',
-      description: [
-        {type: 'p', content: "This is the map for the Friends of Bowker Creek and Peninsula Streams Society '1000 Rain Gardens' mapping project!"},
-        {type: 'link', content: 'Visit and contribute to this project at OpenGreenMap.', url: 'https://new.opengreenmap.org/browse/sites?map=63e6939eabcc260100514352'},
-      ],
-      data: require('./geojson/1000RainGardens.geojson'),
-      shape: 'point',
-      symbology: 'classified',
-      styleMap: styleMap_RainGardens,
-      options: {
-        pointToLayer: (f,l) => {
-          const ogmIcon = icon({
-            iconUrl: `https://new.opengreenmap.org/api-v1/icons/${f.properties.icons[0]}/image/value`,
-            iconSize: [48,48],
-          })
-
-          return f.properties?.icons[0] ? marker(l, {icon: ogmIcon}) : null;
-        },
-        onEachFeature: (f,l) => {
-          const imageURL = f.properties.pictures[0]
-            ? `https://new.opengreenmap.org/api-v1/pictures/${f.properties.pictures[0]}/picture/sm`
-            : null;
-          
-          l.bindPopup(mapPopupContent(
-              f.properties.name,
-              '',
-              `https://new.opengreenmap.org/browse/sites/${f.properties._id}`,
-              'Show More at OpenGreenMap.org',
-              imageURL
-            ), {offset: point(0,-6)});
-        }
-      },
-    },
-  ],
+  }
 };
 export default Flooding;
