+++ src/data/Footprint.jsx
@@ -1,42 +1,3 @@
-import L from 'leaflet';
-import 'leaflet-layerjson';
-import {
-  getStyleMapProperty,
-  getStyleMapKeyFromContinuousValue,
-  mapPopupContent,
-  pointToIcon,
-  pointToIconByProperty,
-} from './LeafletStyleHelpers';
-// SVG
-import { GHGCarIcon } from '../components/EmissionsIcon';
-import { ReactComponent as WaterIcon } from './svg/WaterTransport.svg';
-import { ReactComponent as AirportIcon } from './svg/Airport.svg';
-import { ReactComponent as EVIcon } from './svg/CarEV.svg';
-
-const styleMap_GHG = new Map([
-  [0,    {fillColor: 'rgb(7,72,174)', color: 'rgb(130,130,130)', legendText: 'Reduction'}],
-  [100,  {fillColor: 'rgb(248,56,8)', color: 'rgb(130,130,130)', legendText: 'Increase'}]
-]);
-
-const styleMap_CRDBikeMap = new Map([
-  ["T1" , {legendText: 'Paved Multi-Use Trails / Protected Bike Lanes', color: 'rgba(120, 162, 47, 255)'}],
-  ["T1d", {legendText: 'Unpaved Multi-Use Trails', color: 'rgba(137, 90, 68, 255)'}],
-  ["BL1", {legendText: 'Bike Lanes / Road Shoulders', color: 'rgba(0, 139, 176, 255)'}],
-  ["SR1", {legendText: 'Shared Streets', color: 'rgba(236, 136, 29, 255)'}],
-  ["DC" , {legendText: 'Difficult Connections', color: 'rgba(211, 18, 69, 255)'}]
-]);
-
-const styleMap_TransportHubs = new Map([
-  ['Ferry',   {icon: <WaterIcon />, fill: 'rgb(13,101,149)', legendText: 'Ferry Terminal'}],
-  ['Airport', {icon: <AirportIcon />, fill: 'rgb(100,186,50)', legendText: 'Airport'}],
-])
-
-const styleMap_OpenChargeMap = new Map([
-  ['1', {icon: <EVIcon />, fill: 'rgb(7,72,174)', legendText: 'Level 1 (<2 kW)'}],
-  ['2', {icon: <EVIcon />, fill: 'rgb(130,173,70)', legendText: 'Level 2 (2 - 40 kW)'}],
-  ['3', {icon: <EVIcon />, fill: 'rgb(253,106,33)', legendText: 'Level 3 (>40 kW)', subText: ['Fast Charger']}],
-])
-
 const Footprint = {
   title: 'Light Footprint?',
   question: 'Self-propelled "Light Footprint" Travel',
@@ -80,201 +41,5 @@ const Footprint = {
       },
     ],
   },
-  mapData: [
-    {
-      title: 'Greenhouse Gas Emissions (On-road Transportation)',
-      description: [
-        {type:'p', content:'The Capital Regional District (CRD) has established 2007 as a baseline year where the greenhouse gas (GHG) emissions were calculated. The most recent reporting year was 2020, and this map layer shows which CRD member governments have reduced or increased their on-road transportation 2020 emissions compared to 2007.'},
-        {type:'p', content:'The GHGs released to the atmosphere to be reported in the Transportation Sector are those from combustion of fuels in journeys by on-road, railway, waterborne navigation, aviation, and off-road. GHG emissions are produced directly by the combustion of fuel, and indirectly using grid-supplied electricity.'},
-        {type:'p', content:'Learn more about the local sources of GHGs by reading the reports here:'},
-        {
-          type:'link',
-          content:'Stantec Consulting Ltd., 2021. Capital Regional District - Municipalities and Electoral Areas: 2007 Base Year and 2020 Reporting Year Energy & GHG Emissions Inventory. Capital Regional District, Victoria.',
-          url:'https://www.crd.bc.ca/about/data/climate-change'
-        }
-      ],
-      data: require('./geojson/CRDLocalGovGHG.geojson'),
-      format: 'polygon',
-      symbology: 'classified',
-      styleMap: styleMap_GHG,
-      legendTitle: 'Emissions in 2020 compared to 2007',
-      options: {
-        style: function (feature) {
-          const baseStyle = {
-            stroke: true,
-            opacity: 0.5,
-            color: 'rgb(130,130,130)',
-            dashArray: '',
-            lineCap: 'butt',
-            lineJoin: 'miter',
-            weight: 3,
-            fill: true,
-            fillOpacity: 0.3,
-            interactive: true
-          }
-          return {
-            ...baseStyle,
-            fillColor: getStyleMapProperty(
-                'fillColor',
-                getStyleMapKeyFromContinuousValue(
-                  feature.properties.OnRoadTransportationChange,
-                  styleMap_GHG),
-                styleMap_GHG
-              ),
-          }
-        },
-        // pointToLayer only for point features (City Halls)
-        pointToLayer: (f,l) => pointToIcon(l, {
-          fill: null,
-          stroke: null,
-          icon: <GHGCarIcon percentGHG={f.properties.OnRoadTransportationChange} />
-        }, Math.min(Math.abs(f.properties.OnRoadTransportationChange) + 50, 100), null),
-        onEachFeature: (f,l) => {
-          l.bindPopup(mapPopupContent(
-              f.properties.LocalGov,
-              `${Math.abs(f.properties.OnRoadTransportationChange).toFixed(1)}% ${f.properties.OnRoadTransportationChange > 0 ? 'increase' : 'reduction'} in on-road transportation GHG emissions in 2020 compared to 2007 levels.`
-            ), {offset: [0,8]});
-          l.on({
-            mouseover: (e) => {
-              if (typeof e.target.setStyle === 'function')
-                  e.target.setStyle({ fillOpacity: 0.6 })
-              },
-            mouseout: (e) => {
-              if (typeof e.target.setStyle === 'function')
-                e.target.setStyle({ fillOpacity: 0.3 })
-              }
-          });
-        }
-      }
-    },
-    {
-      title: 'CRD Bike Map',
-      description: 'The CRD Bike Map represents the cycling network throughout the region as informed by the Pedestrian and Cycling Master Plan, a key part of the Regional Transportation Plan. Encouraging cycling contributes to the vision for our communities as established in our Regional Growth Strategy.',
-      data: require('./geojson/CRDBikeMap.geojson'),
-      shape: 'line',
-      symbology: 'classified',
-      styleMap: styleMap_CRDBikeMap,
-      options: {
-        style: function (feature) {
-          const baseStyle = {
-            opacity: 0.9,
-            stroke: true,
-            fill: false,
-            interactive: true,
-            weight: 3,
-          }
-          return {
-            ...baseStyle,
-            color: getStyleMapProperty(
-                'color',
-                feature.properties.BkMapCarto,
-                styleMap_CRDBikeMap
-              ),
-          }
-        },
-        onEachFeature: (f,l) => {
-          l.bindPopup(
-            mapPopupContent(
-              'CRD Bike Map',
-              f.properties.Label,
-              ),
-            {offset: [0,8]});
-        }
-      }
-    },
-    {
-      title: 'BC Transit Bus Routes',
-      description: 'The bus transportation routes offered by BC Transit. Some routes shown may be alternate or infrequently operated schedules.',
-      data: require('./geojson/BCTransitRoutes.geojson'),
-      shape: 'line',
-      symbology: 'single',
-      options: {
-        style: {
-          stroke: true,
-          opacity: 0.9,
-          fill: false,
-          interactive: true,
-          color: 'rgb(18, 144, 201)',
-          weight: 3,
-        },
-        onEachFeature: (f,l) => {
-          l.bindPopup(
-            mapPopupContent(
-              'Route ' + f.properties.Route,
-              f.properties.RouteName,
-              ),
-            {offset: [0,8]});
-        }
-      }
-    },
-    {
-      title: 'EV Charging Stations',
-      description: [
-        {type: 'p', content: 'Charging station data provided by Open Charge Map.'},
-        {type: 'link', content: 'View detailed charging station information, contribute to the map, and learn more about the project here.', url: 'https://openchargemap.org/site'}
-      ],
-      legendTitle: 'Charging Level (Speed)',
-      layer: L.layerJSON({
-        url: "https://api.openchargemap.io/v3/poi/?output=geojson&countrycode=CA&usagetypeid=1&statustypeid=50&maxresults=3000&compact=false&camelcase=true&verbose=false&includecomments=true&key=ca66f44f-646d-4f64-98ec-639b924d2839&boundingbox=({lat1}%2C{lon1})%2C({lat2}%2C{lon2})",
-        propertyItems: 'features',
-        propertyTitle: 'properties.name',
-        propertyLoc: 'geometry.coordinates',
-        updateOutBounds: '',
-        locAsGeoJSON: true,
-        precision: 5,
-        caching: true,
-        cacheId: function(data, latlng) {
-          return data.id || latlng.toString();
-        },
-        dataToMarker: function(data, latlng) {
-          return pointToIconByProperty(
-            data,
-            latlng,
-            "level",
-            styleMap_OpenChargeMap)
-            .bindPopup(
-              mapPopupContent(
-                data.properties.name,
-                data.properties.connectionType
-              )
-            )
-        },
-      }),
-      shape: 'point',
-      symbology: 'classified',
-      styleMap: styleMap_OpenChargeMap,
-    },
-    {
-      title: 'Major Transportation Hubs',
-      description: '',
-      data: require('./geojson/GHGTransportHubs.geojson'),
-      shape: 'point',
-      symbology: 'classified',
-      styleMap: styleMap_TransportHubs,
-      options: {
-        pointToLayer: (f,l) => pointToIconByProperty(
-          f,
-          l,
-          "Icon",
-          styleMap_TransportHubs
-        ),
-        onEachFeature: (f,l) => {
-          l.bindPopup(
-            (f.properties.Name === "BC Ferries"
-              ? `<img src="${require('./png/BCFerriesGHGEmissions.png')}" /><img src="${require('./png/Top6EmittingBCFerriesRoutes.png')}" />`
-              : `<img src="${require('./png/VictoriaInternationalAirportGHGEmissions.png')}" />`
-            ),
-            {
-              offset: [0,-2],
-              autoPan: false,
-              autoClose: false,
-              minWidth: 400,
-              maxWidth: 400
-            });
-          l.on('add', (e) => e.target.openPopup());
-        }
-      }
-    }
-  ],
 };
 export default Footprint;
