+++ src/data/HelpWithChange.jsx
@@ -1,18 +1,12 @@
-// GeoJSON
-// SVG
-// PNG
-import image from './png/HelpWithChange.png';
-
 const HelpWithChange = {
   title: "Help with Change?",
   question: "Who can I contact to express my thoughts on these issues?",
-  image: image,
+  image: require('./png/HelpWithChange.png'),
   description: [
   ],
   act: {
     initiatives: [
     ],
   },
-  mapData: [],
 };
 export default HelpWithChange;
