+++ src/data/Naturehood.jsx
@@ -1,59 +1,3 @@
-import { point } from 'leaflet';
-import { getStyleMapProperty, pointToIcon, mapPopupContent, ogmIconLink } from './LeafletStyleHelpers';
-// GeoJSON
-import Pollinators from './geojson/Pollinators.geojson';
-import StewardshipGroups from './geojson/StewardshipGroups.geojson';
-// SVG
-
-const styleMap_Pollinators = new Map([
-  ['5c311b3537407e1f05ac3946', { src: ogmIconLink('5c311b3537407e1f05ac3946'), legendText: "SDG 3 - Good Health and Well Being" }],
-  ['5c312b3c37407e1f05ac3950', { src: ogmIconLink('5c312b3c37407e1f05ac3950'), legendText: "SDG 13 - Climate Action" }],
-  ['5c312ca837407e1f05ac3952', { src: ogmIconLink('5c312ca837407e1f05ac3952'), legendText: "SDG 15 - Life on Land" }],
-  ['5ca7bfc0ecd8490100cab980', { src: ogmIconLink('5ca7bfc0ecd8490100cab980'), legendText: "Organic or Local Food" }],
-  ['5ca7bfe1ecd8490100cab9d5', { src: ogmIconLink('5ca7bfe1ecd8490100cab9d5'), legendText: "Green School" }],
-  ['5ca7bfe8ecd8490100cab9ed', { src: ogmIconLink('5ca7bfe8ecd8490100cab9ed'), legendText: "Natural Corridor or Greenway" }],
-  ['5ca7bfe8ecd8490100cab9ef', { src: ogmIconLink('5ca7bfe8ecd8490100cab9ef'), legendText: "Shaded Boulevard" }],
-  ['5ca7bfe9ecd8490100cab9f0', { src: ogmIconLink('5ca7bfe9ecd8490100cab9f0'), legendText: "Drinking Water Source" }],
-  ['5ca7bfe9ecd8490100cab9f1', { src: ogmIconLink('5ca7bfe9ecd8490100cab9f1'), legendText: "Future Redevelopment Space" }],
-  ['5ca7bfe9ecd8490100cab9f2', { src: ogmIconLink('5ca7bfe9ecd8490100cab9f2'), legendText: "Eco Design or Planning Feature" }],
-  ['5ca7bfeaecd8490100cab9f3', { src: ogmIconLink('5ca7bfeaecd8490100cab9f3'), legendText: "Waterfront or Riverside Park" }],
-  ['5ca7bfeaecd8490100cab9f5', { src: ogmIconLink('5ca7bfeaecd8490100cab9f5'), legendText: "Eco Landscaping" }],
-  ['5ca7bfececd8490100cab9fa', { src: ogmIconLink('5ca7bfececd8490100cab9fa'), legendText: "Park or Recreation Area" }],
-  ['5ca7bff0ecd8490100caba08', { src: ogmIconLink('5ca7bff0ecd8490100caba08'), legendText: "Insect Watching Site" }],
-  ['5ca7bff2ecd8490100caba0c', { src: ogmIconLink('5ca7bff2ecd8490100caba0c'), legendText: "Bird and Wildlife Watching" }],
-  ['5ca7bff2ecd8490100caba0d', { src: ogmIconLink('5ca7bff2ecd8490100caba0d'), legendText: "Farm Animals" }],
-  ['5ca7bff4ecd8490100caba0f', { src: ogmIconLink('5ca7bff4ecd8490100caba0f'), legendText: "Significant Habitat" }],
-  ['5ca7bff4ecd8490100caba10', { src: ogmIconLink('5ca7bff4ecd8490100caba10'), legendText: "Spring Blossoms" }],
-  ['5ca7bff5ecd8490100caba11', { src: ogmIconLink('5ca7bff5ecd8490100caba11'), legendText: "Native Forest or Plants" }],
-  ['5ca7bff6ecd8490100caba12', { src: ogmIconLink('5ca7bff6ecd8490100caba12'), legendText: "Community Garden" }],
-  ['5ca7bff8ecd8490100caba13', { src: ogmIconLink('5ca7bff8ecd8490100caba13'), legendText: "Food Gathering Site" }],
-  ['5ca7bff9ecd8490100caba14', { src: ogmIconLink('5ca7bff9ecd8490100caba14'), legendText: "Special Tree" }],
-  ['5ca7bff9ecd8490100caba15', { src: ogmIconLink('5ca7bff9ecd8490100caba15'), legendText: "Rapidly Renewable Plants" }],
-  ['5ca7bff9ecd8490100caba16', { src: ogmIconLink('5ca7bff9ecd8490100caba16'), legendText: "Autumn Leaves" }],
-  ['5ca7bff9ecd8490100caba17', { src: ogmIconLink('5ca7bff9ecd8490100caba17'), legendText: "Public Forest or Natural Area" }],
-  ['5ca7bffaecd8490100caba18', { src: ogmIconLink('5ca7bffaecd8490100caba18'), legendText: "Garden" }],
-]);
-
-const styleMap_SensitiveEcosystems = new Map([
-  ['Cliff', { fillColor: 'rgb(156,156,156)' }],
-  ['Coastal Bluff', { fillColor: 'rgb(209,165,255)' }],
-  ['Estuarine', { fillColor: 'rgb(214,133,137)' }],
-  ['Fresh Water', { fillColor: 'rgb(101,119,205)' }],
-  ['Fringe Forest', { fillColor: 'rgb(0,115,76)' }],
-  ['Herbaceous', { fillColor: 'rgb(245,202,123)' }],
-  ['Terrestrial Herbaceous', { fillColor: 'rgb(255,201,28)', legendText: 'Herbaceous (Terrestrial)' }],
-  ['Intertidal', { fillColor: 'rgb(168,112,0)' }],
-  ['Mature Forest', { fillColor: 'rgb(255,236,175)' }],
-  ['Older Forest', { fillColor: 'rgb(156,54,13)' }],
-  ['Older Second Growth', { fillColor: 'rgb(235,94,42)' }],
-  ['Riparian', { fillColor: 'rgb(156,207,0)' }],
-  ['Seasonally Flooded Agricultural', { fillColor: 'rgb(191,255,189)' }],
-  ['Sparsely Vegetated', { fillColor: 'rgb(158,187,215)' }],
-  ['Wetland', { fillColor: 'rgb(0,156,92)' }],
-  ['Woodland', { fillColor: 'rgb(255,158,212)' }],
-  ['Non-Sensitive', { fillColor: 'rgb(255,255,189)' }],
-]);
-
 const Naturehood = {
   title: "Neighbourhood to Naturehood?",
   question: "What is the Nature of this place?",
@@ -143,123 +87,6 @@ const Naturehood = {
         image: require('./png/Placeholder.png')
       }
     ],
-  },
-  mapData: [
-    {
-      title: 'Sensitive Ecosystems',
-      description: [
-        {type:'p', content:'The Sensitive Ecosystems Inventory (SEI) of East Vancouver Island project has been responsible for tracking sensitive ecosystems for many regions within the Capital Regional District (CRD). Their data is what you see on the map. Their research identifies different sensitive ecosystems and allows us to understand better how these ecosystems are at risk. Some of the most significant factors responsible for the degradation and ecosystem damage are human usage and the introduction of invasive species. These ecosystems are shrinking, often due to development or land use changes. The areas on this map do not include private lands, provincial & federal lands, or Indigenous reservations.'},
-        {type:'p', content:'It is important to understand that each of these ecosystems plays a crucial role in the ecological well-being of this area. Damage to one sensitive ecosystem will affect another. These ecosystems in this area of the coast are known for their biodiversity.'},
-        {type:'p', content:'Learn more at the SEI project page:'},
-        {
-          type:'link',
-          content:'"East Vancouver Island & the Gulf Islands," Terrestrial Ecosystem Information. Government of British Columbia, Victoria, Canada. Accessed June 1, 2023.',
-          url:'https://www2.gov.bc.ca/gov/content/environment/plants-animals-ecosystems/ecosystems/search-ecosystem-info/east-vancouver-island-gulf-islands'
-        }
-      ],
-      data: require('./geojson/SensitiveEcosystems.geojson'),
-      format: 'polygon',
-      symbology: 'classified',
-      styleMap: styleMap_SensitiveEcosystems,
-      options: {
-        style: function (feature) {
-          const baseStyle = {
-            stroke: true,
-            opacity: 0.5,
-            color: 'rgb(130,130,130)',
-            dashArray: '',
-            lineCap: 'round',
-            lineJoin: 'round',
-            weight: 2,
-            fill: true,
-            fillOpacity: 0.7,
-            interactive: true
-          }
-          return {
-            ...baseStyle,
-            fillColor: getStyleMapProperty(
-                'fillColor',
-                feature.properties.Ecosystem,
-                styleMap_SensitiveEcosystems
-              ),
-          }
-        },
-        onEachFeature: (f,l) => {
-          l.bindPopup(mapPopupContent(
-              'Sensitive Ecosystem: ' + f.properties.Ecosystem,
-              f.properties.Location
-            ), {offset: [0,8]});
-          l.on({
-            mouseover: (e) => {
-              if (typeof e.target.setStyle === 'function')
-                  e.target.setStyle({ fillOpacity: 1.0 })
-              },
-            mouseout: (e) => {
-              if (typeof e.target.setStyle === 'function')
-                e.target.setStyle({ fillOpacity: 0.7 })
-              }
-          });
-        }
-      }
-    },
-    {
-      title: "G.V.S.P. Pollinators Map",
-      description: [
-        {type:'p', content: 'Greater Victoria / Saanich Peninsula Pollinator Map. Where are the natural areas and gardens that support pollinators in this region?'},
-        {type:'link', content: 'Learn more and contribute to this project at OpenGreenMap.', url: 'https://new.opengreenmap.org/browse/sites?map=621fe03eba477f0100989741'}
-      ],
-      data: Pollinators,
-      shape: 'point',
-      symbology: 'classified',
-      styleMap: styleMap_Pollinators,
-      options: {
-        pointToLayer: (f,l) => pointToIcon(l, {icon: 
-        f.properties.icon ? (
-          <img
-            width='100%'
-            height='100%'
-            src={ogmIconLink(f.properties.icon)}
-            alt={styleMap_Pollinators.get(f.properties.icon).legendText}
-          />
-        ) : null }),
-        onEachFeature: (f,l) => {
-          l.bindPopup(mapPopupContent(
-              f.properties.name,
-              f.properties.description[0]
-            ), {offset: point(0,-6)});
-        }
-      }
-    },
-    {
-      title: 'Stewardship Groups',
-      description: [
-        {type: 'p', content: 'Stewardship groups are volunteer-based community organizations that actively support the environmental health and wellbeing of a specific region, park, or watershed through monitoring programs, invasive species management, public outreach, and governmental advocacy.'},
-        {type: 'p', content: 'In addition to those shown on the map, the following groups work throughout the Capital Regional District:'},
-        {type: 'link', content: 'Greater Victoria Green Team', url: 'https://greenteamscanada.ca/our-green-teams/greater-victoria-green-team/'},
-        {type: 'link', content: 'Habitat Acquisition Trust', url: 'https://hat.bc.ca/volunteer'}
-      ],
-      data: StewardshipGroups,
-      shape: 'polygon',
-      symbology: 'single',
-      options: {
-        style: {
-          stroke: true,
-          color: '#477EFF',
-          fill: true,
-          fillOpacity: 0.7,
-          fillColor: 'rgba(0, 120, 160, 255)',
-          interactive: true,
-        },
-        onEachFeature: (f,l) => {
-          l.bindPopup(
-            mapPopupContent(
-              f.properties.Group,
-              `The ${f.properties['Group']} advocates for preserving the ecosystems of the ${f.properties['FocusLoc']} area. Learn more and support their work at:`,
-              f.properties['URL']),
-            {offset: point(0,8)});
-        }
-      }
-    },
-  ],
+  }
 };
 export default Naturehood;
