+++ src/data/layers/HartlandLandfill.jsx
@@ -1,0 +1,40 @@
+import ReactDOMServer from 'react-dom/server';
+import {
+  divIcon,
+  marker,
+  point,
+} from "leaflet";
+import { mapPopupContent } from "../LeafletStyleHelpers";
+import { ReactComponent as HartlandPie } from '../svg/HartlandPie.svg';
+
+const layer = {
+  title: 'Hartland Landfill',
+  description: 'Overall Waste Composition 2022',
+  data: require('../geojson/HartlandPie.geojson'),
+  shape: 'point',
+  symbology: 'single',
+  icon: { icon: <HartlandPie /> },
+  options: {
+    pointToLayer: (f,l) => {
+      return marker(l, {
+        icon: divIcon({
+          className: "",
+          iconSize: [256, 256],
+          iconAnchor: [128, 128],
+          html: ReactDOMServer.renderToString(<HartlandPie />),
+        })
+      })
+    },
+    onEachFeature: (f,l) => {
+      l.bindPopup(mapPopupContent(
+        f.properties['Name'],
+        <HartlandPie />,
+        ), {offset: point(0,-72)});
+    }
+  },
+  questions: [
+    { key: 'circular-economy', group: 'default' }
+  ]
+}
+
+export default layer;
